add dedicated chat tab with chat and tool permissions cards
